grpc: move SecureGrpcClient to new module
For more grpc clients and better error handling.
